arch:	Fix issue when running git-arch during rebase
During a rebase, HEAD is in a deteached state.
This means git-symbolic-ref will not return a branch.

For now, we assume a detached head means we are in a rebase.
Fetch the branch name by parsing
.git/rebase-merge/head-name
as a work-around.

This is hack solution for now. Maybe using the reflog is a better idea?