delorean: Add git-delorean, a script that automatically creates fixup commits.
The script *git-delorean* creates fixup commits to the commits
that are touched by the individual staged chunks.

Example:
 1.a
~1.b
~2.c
 1.d

~ denotes lines with staged changes, the digit a commit, the letters
content on the line.

git-delorean will create two fixup commits, one for commit "2", with
contents "c" and another for "1" with contents "b".

The commits are created in the order traversed by rev-list (which should
be newest first?)