Add git-pp, a script that facilitates (primarily force) pushing
If there are changes to push, git-pp will first print the graph log of
the commits that differ between HEAD and @{upstream}.

After which it will prompt whether to push or not.

It tries to push, and if doesn't succeed, and doesn't detect unexpected
error, then it prompts whether to perform a force push.