delorean:	git-delorean causes merge-conflict with unstaged changes
Running git-delorean in a working tree where a file has both staged and
unstaged changes results in a merge conflict.

Because git commit with a filespec commits all changes (staged/unstaged)
we use the stash as a way to save the working tree in the beginning and
only retain changes staged in the wokring tree.
However, when we later pop the saved stash, this introduces a
merge-conflict.

Pass the --index flag to git-stash pop to avoid this.