Add git-branch-status, similar to git branch -vv
Inspired by the thread
https://stackoverflow.com/questions/7773939/show-git-ahead-and-behind-info-for-all-branches-including-remotes
mainly
https://stackoverflow.com/a/20499690/
and
https://stackoverflow.com/a/7774433/
the "original" branch-status and its fork
https://github.com/bill-auger/git-branch-status
which I've used a lot but have found to be much to slow for my usage.

Initially we just list three columns [ahead x] etc. and the
corresponding origin.

The aim is to replicate the functionality of
https://github.com/bill-auger/git-branch-status
but faster