mark:	Modify script to pure bash, support deletion and multiple branches
- Modify script to use bash variable substitution features.

This may be reverted in the future. The project should strive for
compatability.

- Support deletion of "marks"

It should be convinient to change/delete "marks". At the moment, a
branch can have at most one mark and it starts with #.
For example:
`#a/x`
means branch `x`, marked as `a`.

- Multiple branches can be specified as a whitespace delimeted
  parameters after the mark

For example:
`git mark c '#a/x' '#b/y' 'z'`
will mark branches `x`, `y`, `z` with mark `#c`.
This will also overwrite marks `#a`/`#b` with `#c`.