resolve: HACK	Fix extremly slow script due to network lookup
There's some kind of strange issue with invoking git scripts that use
the /bin/sh (bash in POSIX mode?) under Git for windows. See the issue
https://groups.google.com/g/git-for-windows/c/HxJIyC2hvro

The issue does not appear to occur when using /bin/bash as the
interpreter. Change the interpreter to /bin/bash as a workaround (at
least for now)