resolve:	Accept input from stdin
Adopt git-resolve to not only work on files, but also run on stdin.

Running git-resolve on stdin requires the file to be specified as either
- or /dev/stdin as there's no reliable way to check wether a script is
  being feed data from stdin.

Defaulting to stdin was considered, but the typical input is often
"everything relative to current working directory" `.` or "everything in
working tree" `:/`.

Methods to read from stdin automatically, using tools and methods like
`ifne` but it turned out to not really work. Besides, specifying - or
/dev/stdin is common practise.