resolve:	Add (z)diff3 conflict style support
There are non-default style conflict markers, see
https://git-scm.com/docs/merge-config#Documentation/merge-config.txt-mergeconflictStyle

The default is "merge", which shows a <<<<<<< conflict marker, changes
made by one side, a ======= marker, changes made by the other side, and
then a >>>>>>> marker.

An alternate style, "diff3", adds a ||||||| marker and the original text
before the ======= marker.

There's also zdiff3, see
https://github.com/git/git/commit/4496526f80b3e4952036550b279eff8d1babd60a

To accommodate for these non-default varieties we modify the sed script a
little.

For both, delete <<<<<<< and >>>>>>> lines as well as lines in-between
patterns ||||||| and ======= and also ======= lines
 1
 2
 3
 4
 A
-<<<<<<
 B
 C
 D
-||||||
-5
-6
-======
 X
 C
 Y
->>>>>>
 E
 7
 8
 9

For ours, delete <<<<<<< and as well as lines in-between
patterns ||||||| or  ======= and the next >>>>>>>
 1
 2
 3
 4
 A
-<<<<<<
 B
 C
 D
-||||||
-5
-6
-======
-X
-C
-Y
->>>>>>
 E
 7
 8
 9

For theirs, delete lines in-between <<<<<<< and =======
as well as >>>>>>> lines. Note that there's no difference in
implementation here.
 1
 2
 3
 4
 A
-<<<<<<
-B
-C
-D
-||||||
-5
-6
-======
 X
 C
 Y
->>>>>>
 E
 7
 8
 9