git-ignore:	Ignore case (fix flag) and sort uniquely
The git-ignore.sh script has been updated to ignore case when sorting
and to remove duplicate entries.

Previously, the incorrect flag "-i" was used to ignore the case, this
has been fixed to "--ignore-case" (-f may be need for POSIX).

Also, if adding a file which was already ignored, it was previously
added again.