Makefile:	add help target to display helpful information
There used to be a PHONY target `ls`, which listed basic debug information.
Which programs are installed, which links are installed and the
installation directory.

We also add an awk file which parses the Makefile and lists
"available-targets".

This awk file parses the Makefile to find targets which are after a
documentation snippet. Such a snippet is made up of lines which are all
prepended with 2 hashes `^##`.

Add such documentation snippets to targets which are relevant to most users.

This is not an exhaustive list of available targets, rather it's a list
of targets the user is likely to be interested in running.

It's intended to be useful, but not something the user should depend on.
It may be incorrect and when in doubt, consult the Makefile.