make: Change bin/ and $(INSTALL_DIR)/ to order-only prerequisites
These pre-requisites (folders) are needed to build targets which reside
in those folders.

Making the order-only, using |, means they are only built if they are
missing but if they are newer, they will not trigger re-building target.

They are also excluded from things like the built-in `$^` variable.

Also, omit the slash `/` from their name.

Not sure if this is preferable or not. I like reading
`$(INSTALL_DIR)/%.sh` instead of `$(INSTALL_DIR)%.sh` because it's
clearer that it's a directory.

There may be other consideration that should be done, so maybe we will
revisit this decision in the future.